demo: on restart, maintain HUD if up

+++ src/demo/demo.c
@@ -439,18 +439,27 @@ int main(int argc, char** argv){
       nanosleep(&demodelay, NULL);
     }
   }
+  // if we're restarted with the HUD up, we want it to seamlessly restart
+  bool start_with_hud = false;
   do{
     restart_demos = false;
     interrupted = false;
     notcurses_drop_planes(nc);
+    if(start_with_hud){
+      if(!hud_create(nc)){
+        goto err;
+      }
+    }
     if(menu_create(nc) == NULL){
       goto err;
     }
     if(ext_demos(nc, spec, ignore_failures) == NULL){
       goto err;
     }
-    if(hud_destroy()){ // destroy here since notcurses_drop_planes will kill it
-      goto err;
+    if( (start_with_hud = !!hud) ){
+      if(hud_destroy()){ // destroy here since notcurses_drop_planes will kill it
+        goto err;
+      }
     }
   }while(restart_demos);
   if(stop_input()){
